# services/policy/openapi.yaml
openapi: 3.1.0
info:
  title: 契約サービス (Policy Service)
  description: 成立した保険契約を管理するAPI。証券番号・保険金額・保険料・契約ステータスを保持します。
  version: 1.0.0
paths:
  /health:
    get:
      tags:
      - health
      summary: ヘルスチェック
      operationId: health_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /policies:
    get:
      tags:
      - policies
      summary: 契約一覧の取得
      operationId: list_policies_policies_get
      parameters:
      - name: customer_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 顧客IDで絞り込み
          title: Customer Id
        description: 顧客IDで絞り込み
      - name: product_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 商品IDで絞り込み
          title: Product Id
        description: 商品IDで絞り込み
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: ステータスで絞り込み
          title: Status
        description: ステータスで絞り込み
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - policies
      summary: 契約の新規登録
      operationId: create_policy_policies_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyBase'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /policies/{policy_id}:
    get:
      tags:
      - policies
      summary: 契約の取得
      operationId: get_policy_policies__policy_id__get
      parameters:
      - name: policy_id
        in: path
        required: true
        schema:
          type: string
          title: Policy Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - policies
      summary: 契約の更新
      operationId: update_policy_policies__policy_id__put
      parameters:
      - name: policy_id
        in: path
        required: true
        schema:
          type: string
          title: Policy Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyBase'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - policies
      summary: 契約の削除
      operationId: delete_policy_policies__policy_id__delete
      parameters:
      - name: policy_id
        in: path
        required: true
        schema:
          type: string
          title: Policy Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    Beneficiary:
      properties:
        name:
          type: string
          title: Name
          description: 受取人氏名
        relationship:
          type: string
          title: Relationship
          description: 続柄
      type: object
      required:
      - name
      - relationship
      title: Beneficiary
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    InsuredPet:
      properties:
        species:
          type: string
          title: Species
          description: 種別
        breed:
          type: string
          title: Breed
          description: 品種
        name:
          type: string
          title: Name
          description: ペットの名前
        age:
          type: integer
          title: Age
          description: ペットの年齢
      type: object
      required:
      - species
      - breed
      - name
      - age
      title: InsuredPet
    Policy:
      properties:
        policy_number:
          type: string
          title: Policy Number
          description: 証券番号
          examples:
          - 2025-FIRE-000001
        application_id:
          type: string
          title: Application Id
          description: 由来の申込ID
          examples:
          - APP-000001
        customer_id:
          type: string
          title: Customer Id
          description: 契約者(顧客ID)
          examples:
          - CUS-000001
        product_id:
          type: string
          title: Product Id
          description: 契約商品(商品ID)
          examples:
          - PRD-001
        contract_date:
          type: string
          title: Contract Date
          description: 契約日(YYYY-MM-DD)
        effective_date:
          type: string
          title: Effective Date
          description: 保険始期日(YYYY-MM-DD)
        expiry_date:
          anyOf:
          - type: string
          - type: 'null'
          title: Expiry Date
          description: 保険終期日(YYYY-MM-DD)
        sum_insured:
          type: integer
          title: Sum Insured
          description: 契約保険金額(円)
        premium_amount:
          type: integer
          title: Premium Amount
          description: 保険料(円)
        premium_payment_cycle:
          type: string
          title: Premium Payment Cycle
          description: 払込周期
          examples:
          - 月払
        payment_method:
          type: string
          title: Payment Method
          description: 支払方法
          examples:
          - 口座振替
        status:
          type: string
          title: Status
          description: 契約ステータス
          examples:
          - 有効
        beneficiary:
          anyOf:
          - $ref: '#/components/schemas/Beneficiary'
          - type: 'null'
          description: 受取人情報
        insured_pet:
          anyOf:
          - $ref: '#/components/schemas/InsuredPet'
          - type: 'null'
          description: 被保険動物(ペット保険のみ)
        riders:
          items:
            type: string
          type: array
          title: Riders
          description: 付帯特約
        policy_id:
          type: string
          title: Policy Id
          description: 契約ID(POL-NNNNNN)
          examples:
          - POL-000001
        created_at:
          type: string
          title: Created At
        updated_at:
          type: string
          title: Updated At
      type: object
      required:
      - policy_number
      - application_id
      - customer_id
      - product_id
      - contract_date
      - effective_date
      - sum_insured
      - premium_amount
      - premium_payment_cycle
      - payment_method
      - status
      - policy_id
      - created_at
      - updated_at
      title: Policy
    PolicyBase:
      properties:
        policy_number:
          type: string
          title: Policy Number
          description: 証券番号
          examples:
          - 2025-FIRE-000001
        application_id:
          type: string
          title: Application Id
          description: 由来の申込ID
          examples:
          - APP-000001
        customer_id:
          type: string
          title: Customer Id
          description: 契約者(顧客ID)
          examples:
          - CUS-000001
        product_id:
          type: string
          title: Product Id
          description: 契約商品(商品ID)
          examples:
          - PRD-001
        contract_date:
          type: string
          title: Contract Date
          description: 契約日(YYYY-MM-DD)
        effective_date:
          type: string
          title: Effective Date
          description: 保険始期日(YYYY-MM-DD)
        expiry_date:
          anyOf:
          - type: string
          - type: 'null'
          title: Expiry Date
          description: 保険終期日(YYYY-MM-DD)
        sum_insured:
          type: integer
          title: Sum Insured
          description: 契約保険金額(円)
        premium_amount:
          type: integer
          title: Premium Amount
          description: 保険料(円)
        premium_payment_cycle:
          type: string
          title: Premium Payment Cycle
          description: 払込周期
          examples:
          - 月払
        payment_method:
          type: string
          title: Payment Method
          description: 支払方法
          examples:
          - 口座振替
        status:
          type: string
          title: Status
          description: 契約ステータス
          examples:
          - 有効
        beneficiary:
          anyOf:
          - $ref: '#/components/schemas/Beneficiary'
          - type: 'null'
          description: 受取人情報
        insured_pet:
          anyOf:
          - $ref: '#/components/schemas/InsuredPet'
          - type: 'null'
          description: 被保険動物(ペット保険のみ)
        riders:
          items:
            type: string
          type: array
          title: Riders
          description: 付帯特約
      type: object
      required:
      - policy_number
      - application_id
      - customer_id
      - product_id
      - contract_date
      - effective_date
      - sum_insured
      - premium_amount
      - premium_payment_cycle
      - payment_method
      - status
      title: PolicyBase
    PolicyList:
      properties:
        total:
          type: integer
          title: Total
        items:
          items:
            $ref: '#/components/schemas/Policy'
          type: array
          title: Items
      type: object
      required:
      - total
      - items
      title: PolicyList
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
tags:
- name: policies
  description: 契約の参照・管理
servers:
- url: https://api.example.com/policy
  description: Kong Gateway 経由(本番想定)